==> notebooks/00_build_dir_structure.ipynb <==
# Build Directory Structure for Executing RAG

import sys
current_dir = "/Users/josephtolsma/Documents/dev/yelp_rag"
sys.path.insert(0,current_dir)

import os
from pathlib import Path
import src.config as config
import shutil

**Remove and Replace Intermediate Data**

data_dirs = (config.DATA_DIR_SAMP,config.DATA_DIR_PROC)

for dir in data_dirs:
    dir = Path(dir)
    if dir.exists():
        shutil.rmtree(dir)
    os.mkdir(dir)

**Remove Index and Metadatafiles But Leave Readme Intact**

extensions = (".faiss",".parquet")
for file in Path(config.INDEX_DIR).iterdir():
    if file.is_file():
        if extensions is None or file.suffix in extensions:
            file.unlink()
