==> notebooks/05_retrieval_test.ipynb <==
# Relevant Text Retrieval

import sys
current_dir = "/Users/josephtolsma/Documents/dev/yelp_rag"
sys.path.insert(0,current_dir)

import pandas as pd
import os
import faiss
import src.config as config
from sentence_transformers import SentenceTransformer
from pathlib import Path

pd.set_option("display.max_colwidth",None)

**Pick a sample review id to use for the initial function build**

chunk_df = pd.read_parquet(os.path.join(config.DATA_DIR_PROC,"review_chunks.parquet"))
rid = chunk_df["business_id"].sample(random_state=42).values[0]

**Build retrieval functionality**

def load_embedding_model(model_name,device = "mps") -> SentenceTransformer:
    """
    Load a local embedding model using sentence_transformer library
    """
    return SentenceTransformer(model_name,device)

def convert_similarity_arrays_to_df(score_array,index_array,metadata,metadata_cols):
    """
    Converts raw similarity arrays from FAISS to a dataframe of relevant text chunks.
    Limits returned values to a preconfigured number of chunks per review to avoid duplicated statements.
    @param score_array: similarity scores returned by searching query verbatim in the corpus of reviews
    @param index_array: indicies associated with each chunk returned by the FAISS query search
    @param metadata: metadata dataframe associated with each index, such as business id and (non-encoded) text chunk
    @param metadata_cols: columns in the metadata dataframe relevant to the final result
    @returns dataframe of most similar text chunks and identifiers for each chunk
    """

    chunk_id_col = metadata_cols["chunk_id_col"]
    business_id_col = metadata_cols["business_id_col"]
    restaurant_name_col = metadata_cols["restaurant_name_col"]
    stars_col = metadata_cols["stars_col"]
    review_id_col = metadata_cols["review_id_col"]
    chunk_col = metadata_cols["chunk_col"]

    df = pd.DataFrame({
    "scores":score_array[0],
    "indicies":index_array[0]
    }).assign(
        chunk_id = lambda df: metadata.iloc[df["indicies"]][chunk_id_col].values,
        business_id = lambda df: metadata.iloc[df["indicies"]][business_id_col].values,
        restaurant_name = lambda df: metadata.iloc[df["indicies"]][restaurant_name_col].values,
        stars = lambda df: metadata.iloc[df["indicies"]][stars_col].values,
        review_id = lambda df: metadata.iloc[df["indicies"]][review_id_col].values,
        chunk = lambda df: metadata.iloc[df["indicies"]][chunk_col].values
    )

    df = df.sort_values(by = "scores",ascending=False)
    df = df.groupby("review_id",as_index=False).head(config.MAX_CHUNKS_PER_REVIEW)
    return df


index = faiss.read_index(os.path.join(config.INDEX_DIR,f"{rid}.faiss"))
meta = pd.read_parquet(os.path.join(config.INDEX_DIR,f"{rid}_meta.parquet"))

model = load_embedding_model(config.EMBEDDING_MODEL_NAME,"mps")
query = model.encode(["I want chicken, I want liver, meow mix meow mix please deliver"],convert_to_numpy=True,normalize_embeddings=True)

D,I = index.search(query,k = 3)
metadata_cols = config.METADATA_COLS
convert_similarity_arrays_to_df(D,I,meta,metadata_cols)

**Loop Retrieval Functionality Through All Topics (One Restauarant)**

result_df = pd.DataFrame()

for topic,keywords in config.TOPICS.items():
    query = model.encode([keywords],convert_to_numpy=True,normalize_embeddings=True)
    D,I = index.search(query,k = 12)
    df = convert_similarity_arrays_to_df(D,I,meta,metadata_cols)
    df.insert(0,"topic",topic)
    result_df = pd.concat([result_df,df],axis = 0)

# top results for each topic at the chosen restauarant
result_df.loc[0]

**Loop Retrieval Functionality Through All Topics (All Restaurants)**

faiss_files = sorted(config.INDEX_DIR.glob("*.faiss"))

index_pairs = []
for faiss_path in faiss_files:
    restaurant_id = faiss_path.stem
    meta_path = config.INDEX_DIR / f"{restaurant_id}_meta.parquet"

    if not meta_path.exists():
        raise KeyError(f"Missing metadata for restaurant id {restaurant_id}")
    
    index_pairs.append({
        "restaurant_id":restaurant_id,
        "index_file":faiss_path,
        "metadata_file":meta_path
    })

query_embed = {}
for topic,keywords in config.TOPICS.items():
    query = model.encode([keywords],convert_to_numpy=True,normalize_embeddings=True)
    query_embed.update({topic:query})

result_df = pd.DataFrame()

for file in index_pairs:
    index_file_path = str(file["index_file"])
    meta_file_path = file["metadata_file"]
    index = faiss.read_index(index_file_path)
    meta = pd.read_parquet(meta_file_path)
    for topic in config.TOPICS.keys():
        D,I = index.search(query_embed[topic],k = config.TOP_K_PER_TOPIC)
        df = convert_similarity_arrays_to_df(D,I,meta,config.METADATA_COLS)
        df.insert(0,"topic",topic)
        df = df.iloc[:config.MAX_CHUNKS_PER_TOPIC]
        result_df = pd.concat([result_df,df],axis = 0,ignore_index=True)


pd.concat([
    result_df[result_df["topic"] == "food"].sample(1,random_state=1),
    result_df[result_df["topic"] == "service"].sample(1,random_state=2),
    result_df[result_df["topic"] == "ambiance"].sample(1,random_state=4),
],axis = 0, ignore_index=True
)

result_df.to_parquet(os.path.join(config.DATA_DIR_PROC,"topic_relevant_review_chunks.parquet"))
