==> notebooks/06_summarization_test.ipynb <==
# Summarization Using LLM Wrapper

import sys
current_dir = "/Users/josephtolsma/Documents/dev/yelp_rag"
sys.path.insert(0,current_dir)

import pandas as pd
import os
import requests
import src.config as config

pd.set_option("display.max_colwidth",40)

relevant_chunk_df = pd.read_parquet(os.path.join(config.DATA_DIR_PROC,"topic_relevant_review_chunks.parquet"))

relevant_chunk_df.head(3)

pd.set_option("display.max_colwidth",None)

**Build Content for Summarization for One Restaurant**

rid = relevant_chunk_df.sample(random_state=42)["business_id"].values[0]

chunk_test_df = relevant_chunk_df[relevant_chunk_df["business_id"] == rid]

topic_context_dict = {}

for topic in config.TOPICS.keys():
    chunk_topic_df = chunk_test_df[chunk_test_df["topic"]==topic]
    context = f"""
    You are analyzing and synthesizing customer reviews for a restaurant called {chunk_test_df["restaurant_name"].values[0]}.

    The following are excerpts from different customer reviews. 
    Each review comes from a different reviewer.

    """
    context+=f"""
    Topic: {topic.upper()}
    """
    i = 0
    for row in chunk_topic_df.iterrows():

        verbatim = f"""
        [{i}] Review ({row[1].stars} stars):
        {row[1].chunk}
        """
        context+=verbatim
        i+=1
    topic_context_dict.update({topic:context})

def call_ollama(
        prompt,
        model,
        temperature=0.2
    ):
    
    url = "http://localhost:11434/api/generate"
    payload = {
        "model":model,
        "prompt":prompt,
        "stream":False,
        "options":{"temperature":temperature}
    }

    resp = requests.post(url,json=payload,timeout=120)
    resp.raise_for_status()
    
    data = resp.json()
    return data["response"].strip()

print(call_ollama(prompt = "say hello.",model = config.OLLAMA_MODEL,temperature=0.2))

summary_dict = {}
for topic in config.TOPICS.keys():
    chunk_topic_df = chunk_test_df[chunk_test_df["topic"]==topic]
    restaurant_name = chunk_test_df["restaurant_name"].values[0]
    context = f"""
    You are analyzing and synthesizing customer reviews for a restaurant called {restaurant_name}, where the main topic is:
    
    Topic: {topic.upper()}

    Using only the reviews included here, provide a concise (2-3 sentence) summary of {restaurant_name}'s performance in the area of {topic}.

    Guidelines:
    - Capture common themes and disagreements
    - Mention both positives and negatives if present
    - Be factual and grounded in the evidence
    - Do not invent details not supported by the excerpts
    - Do not announce the summary ("here is a concise summary..."); just start with summary content

    """
    i = 0
    for row in chunk_topic_df.iterrows():

        verbatim = f"""
        [{i}] Review ({row[1].stars} stars):
        {row[1].chunk}
        """
        context+=verbatim
        i+=1
    
    summary = call_ollama(prompt = context,
                          model = config.OLLAMA_MODEL,
                          temperature=config.TEMPERATURE
                        )
    summary_dict.update({topic:summary})

pd.DataFrame(summary_dict,index = [0])

**Build Content for Summarization for All Restaurants**

restaurant_ids = relevant_chunk_df["business_id"].unique().tolist()

summary_dict = {}
for id in restaurant_ids:
    restaurant_df = relevant_chunk_df[relevant_chunk_df["business_id"] == id]
    restaurant_name = restaurant_df.iloc[0]["restaurant_name"]
    restaurant_summary_dict = {}
    for topic in config.TOPICS.keys():
        topic_df = restaurant_df[restaurant_df["topic"]==topic]
        context = f"""
        You are analyzing and synthesizing customer reviews for a restaurant called {restaurant_name}, where the main topic is:
        
        Topic: {topic.upper()}

        Using only the reviews included here, provide a concise (2-3 sentence) summary of {restaurant_name}'s performance in the area of {topic}.

        Guidelines:
        - Capture common themes, impressions, and disagreements
        - Be factual and grounded in the evidence; do not invent details not supported by the excerpts
        - Begin by characterizing the group of reviews as "positive", "negative", or "mixed" 

        Output to avoid:
        - Content unrelated to the topic (for example: "noise levels" are ambiance, not food)
        - Review indicies (for example, "review 3"); those are not user-facing
        - Announcing the summary ("here is a concise summary..."); just start with summary content
        - Use of generic filler adjectives (e.g., cozy, delicious, amazing) unless the excerpts explicitly support them

        Important: if the excerpts don’t describe a topic clearly, say ‘<topic> details are limited in the provided reviews.’”

        """
        i = 0
        for row in chunk_topic_df.iterrows():

            verbatim = f"""
            [{i}] Review ({row[1].stars} stars):
            {row[1].chunk}
            """
            context+=verbatim
            i+=1
        
        summary = call_ollama(prompt = context,
                            model = config.OLLAMA_MODEL,
                            temperature=config.TEMPERATURE
                            ).replace("\n"," ")
        restaurant_summary_dict.update({topic:summary})
    summary_dict.update({restaurant_name:restaurant_summary_dict})

pd.DataFrame(summary_dict).T
