==> notebooks/07_yelp_rag_runner_test.ipynb <==
import sys
current_dir = "/Users/josephtolsma/Documents/dev/yelp_rag"
sys.path.insert(0,current_dir)

import src.config as config
import os
from src.data_io import ImportYelpReviewText
from src.cleaning import CleanChunkYelpReviews

io = ImportYelpReviewText(
    os.path.join(config.DATA_DIR_RAW,"yelp_academic_dataset_review.json"),
    os.path.join(config.DATA_DIR_RAW,"yelp_academic_dataset_business.json"),
    os.path.join(config.DATA_DIR_SAMP,"reviews_samples.csv"),
    os.path.join(config.DATA_DIR_SAMP,"reviews_df.csv")
)
io.generate_final_restaurant_list_for_rag()

cl = CleanChunkYelpReviews(
    os.path.join(config.DATA_DIR_SAMP,"reviews_df.csv"),
    os.path.join(config.DATA_DIR_PROC,"review_chunks.parquet")
)

cl.clean_chunk_export()

import pandas as pd
pd.read_csv(os.path.join(config.DATA_DIR_SAMP,"reviews_df.csv"))
